core: clang: config: add support for global clang exceptions
Oreo requires all modules (except the exceptions) to be
compiled with Clang.

Allow to add anything to global exceptions anywhere.

Since in the device tree often are modules (like HALs) that
should be compiled using GCC in order to retain compatibility
or fix symbol resolving issues, add it to the global exceptions.

Change-Id: I6b335bd44dd67f3ad0def30bfcf6428b6be4f19c